requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
Flask
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tornado==6.1
tzdata==2025.1
tzlocal==5.3
urllib3<2
